feat: Add category image column to Category table
This commit adds a new column `image` to the `Category` table in the database. The column is required and has a data type of VARCHAR(191). This change allows for storing and retrieving images associated with each category. It enhances the functionality of the app by enabling the display of category images in the UI.