feat: Add .env to .gitignore and update npm dependencies
This commit adds the .env file to the .gitignore to prevent it from being tracked in version control. It also updates the npm dependencies by adding "@prisma/client": "^5.18.0" and "mysql2": "^3.11.0" to the package.json file. These changes enhance the project's functionality by ensuring sensitive environment variables are not exposed and keeping the dependencies up to date.